Fix command-line arguments.
Previously the command-line arguments were not entirely clear, which
made debugging difficult.  Also, in the original, running with three
arguments would cause a runtime error due to incorrect number of
parameters.  This patch clears up the usage of the git-commit-notifier
script and ensures that whatever values the user passes will not cause
runtime errors.